+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2025-06-05
+
+### Changed
+
+- Improve tests
+
+### Fixed
+
+- `RedisQueue::size()` do not count available jobs in delayed queue
+- `RedisJob::$attempts` not incremented correctly
+
 ## [1.0.0] - 2025-05-19
 
 No changes were introduced since the previous beta 10 release.